requirements.txt
dj-database-url==0.4.2
Django==2.0.2
django-ckeditor==5.4.0
django-js-asset==1.0.0
djangorestframework==3.7.7
pytz==2018.3
whitenoise==3.3.1
